Summary: Make SPACE_FOR_BBOX be size_t.
ogrgeojsonwritelayer.cpp:90:55: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]


git-svn-id: https://svn.osgeo.org/gdal/trunk@35838 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/geojson/ogr_geojson.h
@@ -175,7 +175,7 @@ class OGRGeoJSONDataSource : public OGRDataSource
 
     virtual void        FlushCache();
 
-    static const int SPACE_FOR_BBOX = 130;
+    static const size_t SPACE_FOR_BBOX = 130;
 
   private:
     //
